% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-cells.r
\name{bootstrap_cells}
\alias{bootstrap_cells}
\title{Bootstrap cell labels for computing confidence intervals}
\usage{
bootstrap_cells(log2counts, grouping_vector, number_bootstrap)
}
\arguments{
\item{log2counts}{log2counts matrix of gene by cells}

\item{grouping_vector}{the grouping vector corresponds to variable of interest}

\item{num_bootstrap}{number of bootstrap samples}
}
\description{
Bootstrap observations (e.g., cells) at each level of the grouping
variable (e.g., individual) to compute confidence interval
for a statistic that has only 1 measurement per individual. This
function was developed for Tung et al. (2006) in comparing
adjusted coefficients of variation computed from three subsets of
cells, each of which is quantified for a LCL.
}
\examples{
bootstrap_cells()

}

