% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-cv-test.r
\name{bootstrap_cv_test}
\alias{bootstrap_cv_test}
\title{Bootstrapping to estimate confidence interval for
individual coefficients of variation}
\usage{
bootstrap_cv_test(log2counts, grouping_vector, anno, number_bootstrap,
  size_bootstrap, output_rda = FALSE, do_parallel = FALSE,
  number_cores = NULL)
}
\arguments{
\item{log2counts}{log2 count matrix of gene by cell.}

\item{number_bootstrap}{Number of boostrapped samples.}

\item{group_cv}{CVs per batch computed use compute_cv().}
}
\description{
For each gene, resample within each individual and compute a
confidence interval of each individual coefficient of variation
based on the boostrapped samples. The method was developed for
Tung et al. (2016) for comparing cell-to-cell heterogeneity between
individuals
}
\examples{
bootstrap_cv_test()

}

