% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-phase-assign.r
\name{cell_phase_assign}
\alias{cell_phase_assign}
\title{Assign cell-phase using an ad hoc approach described 
in Macosco et al. (2015) (http://www.ncbi.nlm.nih.gov/pubmed/26000488)}
\usage{
cell_phase_assign(cell_cycle_genes, expression_matrix)
}
\arguments{
\item{cell_cycle_gene}{a list that is comprised of cell-cycle genes (Ensemble IDs)}

\item{expression}{matrix gene-by-cell matrix. Currently, the method can be applied
to both count data and log-transformed counts.}
}
\value{
score_final_normed Phase-specific score matrix. Scores are standardized 
                           to z-scores across cells and across phases.
}
\description{
This approach was develped for single-cell RNA sequencing data (scRNA-seq)
collected using Drop-seq protocol.
}
\examples{
cell_phase_assign()

}
\seealso{
Other single.cell: \code{\link{density_overlay}}
}

