% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-cv.r
\name{compute_cv}
\alias{compute_cv}
\title{Compute coefficient of variation}
\usage{
compute_cv(log2counts, grouping_vector)
}
\arguments{
\item{log2counts}{log2 count matrix of gene by cell.}

\item{grouping_vector}{Compute per gene CV for cells belonging to each level of
the grouping vector.}
}
\description{
Compute CV across cells belong to each level of the grouping variable. The 
function was developed for Tung et al. (2016) comparing cell-to-cell 
heterogeneity between individuals.
'
}
\examples{
compute_cv()

}

