% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/run_igraphScatter_demo.R
\name{demo_data}
\alias{demo_data}
\title{Import data and prepare objects for shiny app}
\usage{
demo_data(molecules_cv, anno_single, molecules_single_collision, per_person)
}
\description{
Import data and prepare objects for shiny app
}
\examples{
Data already cleaned. Lowly expressed or highly expressed genes were removed.
save(anno_single, molecules_single_collision, molecules_cv,
      file = "../data/igraphScatter-demo.rda")
}

