% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/fisherCorrTest.r
\name{fisherCorrTest}
\alias{fisherCorrTest}
\title{Fisher's test for comparing Pearson's correlation coefficients}
\usage{
fisherCorrTest(cor1, cor2, N1, N2)
}
\arguments{
\item{cor1}{Correlation coefficient}

\item{cor2}{Correlation coefficient}

\item{N1}{Sample size of cor1}

\item{N2}{Sample size of cor2}
}
\description{
This function computes a p-value of the difference between
two Pearson's correlation coefficients
}
\examples{
fisherCorrTest()
}
\keyword{Humanzee}

