% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-lm.r
\name{fit_lm}
\alias{fit_lm}
\title{Differential expression analysis using likelihood
ratio test statistic}
\usage{
fit_lm(exp, species)
}
\arguments{
\item{exp}{Expression set.}

\item{species}{Sample species label (a factor vector).}
}
\description{
This function tests for differential gene expression
using a nested linear model framework.
}
\examples{
fit_lm()
}
\keyword{Humanzee}

