% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get_qval.r
\name{get_qval}
\alias{get_qval}
\title{Multiple hypothesis testings correction}
\usage{
get_qval(pvalues)
}
\arguments{
\item{pvalues}{}
}
\description{
This function uses qvalue package and computes false
discovery rate for each gene.
}
\examples{
get_qval()
}
\keyword{Humanzee}

