% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/gls.series_multiple_designs.r
\name{gls.series_multiple_designs}
\alias{gls.series_multiple_designs}
\title{limma gls.series for varying covariates across genes}
\usage{
gls.series_multiple_designs(M, cov_matrix = NULL, design = NULL,
  ndups = 1, spacing = 1, block = NULL, correlation = NULL,
  weights = NULL, ...)
}
\arguments{
\item{M}{}
}
\description{
This function tests for divergence between two species
in one molecular phenotype at a time.
}
\examples{
M = phos_data
cov_matrix = protein_pheno_data
individual <- as.numeric(str_extract(colnames(phos_data), "[0-9]+"))
design <- model.matrix(~ 1 + as.factor(individual))
block <- block
correlation <- mrho
ndups = 1; weights = NULL; spacing = 1
}
\keyword{Humanzee}

