% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goHeatmapData.R
\name{goHeatmapData}
\alias{goHeatmapData}
\title{Convert Prepare data for visualizing GO terms in a heatmap}
\usage{
goHeatmapData(golist)
}
\arguments{
\item{go_list}{a list that contains summary statistics of the GO terms in GOstats
package. This is used with the GOtest function.}
}
\value{
heatmap_allgos a matrix that contains GO terms, associated annotations,
                       and the p-values of enrichment analysis.
}
\description{
Convert Prepare data for visualizing GO terms in a heatmap
}
\examples{
golist <- list(go_riboGTrna = summary(go_riboGTrna$CC, pvalue = 1),
                go_rnaGTribo = summary(go_rnaGTribo$CC, pvalue = 1),
                go_riboGTpro = summary(go_riboGTpro$CC, pvalue = 1),
                go_proGTribo = summary(go_proGTribo$CC, pvalue = 1) )
heatmap_cc <- goHeatmapData (golist)
}

