% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Humanzee-package.r
\docType{package}
\name{Humanzee-package}
\alias{Humanzee-package}
\title{Gilad lab's tools for comparative analysis of gene expression data}
\description{
This package includes a suite of tools used to study inter-species
variations across multiple molecular phenotypes, using RNAseq,
Ribosome profiling, and quantitative mass spectrometry data.
}
\author{
Chiaowen Joyce Hsiao \email{joyce.hsiao1@gmail.com}

Sidney Wang \email{siddisis@gmail.com}

Yoav Gilad \email{gilad@uchicago.edu}
}

