% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact2way.r
\name{interact2way}
\alias{interact2way}
\title{Interaction model for comparing inter-species divergence}
\usage{
interact2way(eSet, coef = FALSE)
}
\arguments{
\item{eSet}{data set contains the two molecular phenotypes}
}
\description{
This function is a wrapper for an interaction model testing
inter-species differences in divergence between molecular 
phenotypes under a nested linear model framework. We use
likelihood ratio statistic to determine statistical significance
of the difference between inter-species divergence.
}
\examples{
interact2way()
eSet <- eSetRRP.RP.Q.log2[ ,eSetRRP.RP.Q.log2$seqData!="protein" & 
                            eSetRRP.RP.Q.log2$species!="rhesus"]
}
\keyword{Humanzee}

