% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interact2way_full_buffer.r
\name{interact2way_full_buffer}
\alias{interact2way_full_buffer}
\title{Original version of interaction model for comparing inter-speciees divergence}
\usage{
interact2way_full_buffer(eSet)
}
\arguments{
\item{eSet}{data set contains the two molecular phenotypes}
}
\description{
This function is a wrapper for an interaction model testing
inter-species differences in divergence between molecular
phenotypes under a nested linear model framework. We use
likelihood ratio statistic to determine statistical significance
of the difference between inter-species divergence.
}
\examples{
interact2way_full_buffer()
}
\keyword{Humanzee}

