% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedModel2Fit-multiple-design.r
\name{mixedModel2Fit_multiple_design}
\alias{mixedModel2Fit_multiple_design}
\title{Fitting mixed models of one random effect}
\usage{
mixedModel2Fit_multiple_design(design, block, yy)
}
\arguments{
\item{yy}{}
}
\description{
Adapted from statmod package's mixedModel2Fit function. This function
includes the additional feature of fitting a different design matrix
for each gene (row) of the input data matrix.
}
\examples{
mixedModel2Fit_multiple_design()
data_matrix <- pheno_data
design <- model.matrix(~ 1+ as.factor(individual) + protein, 
                       data = pheno_data)
block <- as.factor(pheno_data$biorep)
yy <- pheno_data$phos
}
\keyword{Humanzee}

