% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-cv.r, R/normalize-cv.r
\name{normalize_cv}
\alias{normalize_cv}
\title{Adjust coefficients of variation to remove mean dependency}
\usage{
normalize_cv(group_cv, log2counts, anno)

normalize_cv(group_cv, log2counts, anno)
}
\arguments{
\item{group_cv}{CVs per batch computed use compute_cv().}

\item{log2counts}{log2 count matrix of gene by cell.}

\item{group_cv}{CVs per batch computed use compute_cv().}

\item{log2counts}{log2 count matrix of gene by cell.}
}
\description{
Adjust coefficients of variation (of each group) across genes
to account for the confouding effet of per gene average
expression levels. CVs are computed on the count scale,
and computation of adjusted CV is performed on log10 scale (of CV).
The method was developed for Tung et al. (2016) for comparing 
cell-to-cell heterogeneity between individuals, following 
Kolodziejczyk et al. (2015) (http://dx.doi.org/10.1016/j.stem.2015.09.011).

Adjust coefficients of variation
}
\examples{
normalize_cv()

normalize_cv()

}

