% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute-cv-expressed.r
\name{permute_cv_expressed}
\alias{permute_cv_expressed}
\title{Permute cell labels for computing empirical p-values}
\usage{
permute_cv_expressed(log2counts, grouping_vector, anno, number_permute,
  output_rda = FALSE, do_parallel = FALSE, number_cores = NULL)
}
\arguments{
\item{log2counts}{log2counts matrix of gene by cells.}

\item{grouping_vector}{the grouping vector corresponds to variable of interest.}

\item{anno}{Annotation matrix (cell-by-phenotype).}

\item{number_permute}{number of permuted samples.}
}
\description{
Permute individual labels to compute empirical p-value
for similarity metrics comparing the observed
individual level coefficients of variation (CV). This
function was developed for Tung et al. (2006) in comparing
adjusted coefficients of variation computed from three subsets of
cells, each of which is quantified for a LCL.
}
\examples{
permute_cv_test()

}

