% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute-interaction.r
\name{permute_interact}
\alias{permute_interact}
\title{Wapper function for running interaction model on null data sts}
\usage{
permute_interact(eset_full, datatypes, permute_labels, parallel = FALSE,
  ncores = 4)
}
\arguments{
\item{eset_full}{Expression set.}

\item{datatypes}{Technology type, labels are rna, ribo, or protein.}

\item{permute_labels}{Specificy the phenotype labels that are to be 
shuffled during permutation.}
}
\description{
Wapper function for running interaction model on null data sts
}
\examples{
permute_interact()
}

