% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-density-overlay.r
\name{plot_density_overlay}
\alias{plot_density_overlay}
\title{Overlay density curves using base package}
\usage{
plot_density_overlay(molecules_ENSG, annotation, batches = NULL, which_gene,
  labels = NULL, xlims = NULL, ylims = NULL, gene_symbols = NULL)
}
\arguments{
\item{annotation}{meta-data of the samples.}

\item{which_gene}{Ensemble gene ID of the gene to plot.}

\item{labels}{secondary title right below gene symbol}

\item{gene_symbols}{a gene-by-c(Ensemble ID, gene symbol) matrix. Gene symbols can be
replaced with other gene information of interest.}

\item{expression_matrix}{gene-by-sample matrix with rownames consisting of 
Ensemble gene IDs.}
}
\description{
Overlay density curves using base package
}
\examples{
plot_density_overlay()
}
\seealso{
Other single.cell: \code{\link{cell_phase_assign}}
}

