% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_divergence_scatter.r
\name{plot_divergence_scatter}
\alias{plot_divergence_scatter}
\title{Scatter plot of per gene inter-species divergence estimates}
\usage{
plot_divergence_scatter(xy, ind_points1 = NULL, ind_points2 = NULL, xlab,
  ylab, col_bkpoints = "black", cex_bkpoints = 0.4, lwd_bkpoints = 0.5,
  pch_bkpoints = 1, col_points1 = brocolors("crayons")["Red Orange"],
  col_points2 = brocolors("crayons")["Red Orange"], col_diag = "grey",
  cex_points1 = 0.13, cex_points2 = 0.13, pch_points1 = 20,
  pch_points2 = 20, col_hline = brocolors("crayons")["Pine Green"],
  col_vline = brocolors("crayons")["Pine Green"], lwd_hline = 0.5,
  lwd_vline = 0.5, cex.axis = 0.3, cex.lab = 0.3, xlim = c(-6, 10),
  ylim = c(-6, 10))
}
\arguments{
\item{xy}{}

\item{int_points1}{}
}
\description{
This function is a wrapper for a scatter plot displaying
per gene inter-species divergence estimates
with each gene marked for two divergence categories
by color and shape
}
\examples{
plot_divergence_scatter()
}
\keyword{Humanzee}

