% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot a heatmap of GO terms}
\usage{
plotHeatmap(heatmapData, mar = c(0, 0, 0, 0), oma = c(0, 0, 0, 0),
  keysize = 0.5, labCol = "")
}
\arguments{
\item{heatmapData}{matrix generated by Humanzee::goHeatmapData, which contains
GO terms annotations and p-values of the enrichment analysis.}
}
\value{
hmp heatmap generated by heatmap.2 function.
}
\description{
Plot a heatmap of GO terms
}
\examples{
 hmap <- plotHeatmap(heatmap_allgos,
                     labCol = c("Ribo > mRNA", "RNA > Ribo",
                               "Ribo > Protein", "Protein > Ribo"))
}

