% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/PropTest.r
\name{PropTest}
\alias{PropTest}
\title{Two-proportion z-test}
\usage{
PropTest(prop1, prop2, N1, N2)
}
\arguments{
\item{prop1}{sample proportion}

\item{prop2}{sample proportion}

\item{N1}{Sample size of prop1}

\item{N2}{Sample size of prop2}

\item{propAll}{}
}
\description{
This function compares two indepenent sample proportions.
The null hypothesis is that the two proportions are equal.
}
\examples{
PropCorrTest()
}
\keyword{Humanzee}

