% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_igraphScatter.r
\name{run_igraphScatter}
\alias{run_igraphScatter}
\title{Run Scatter shiny app}
\usage{
run_igraphScatter()
}
\description{
Call igraphScatter. The app was developed for Tung et al. (2016)
to faciliate exploratory data analysis.
}
\examples{
run_igraphScatter()

}

