% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ruv_mixed_model.R
\name{ruv_mixed_model}
\alias{ruv_mixed_model}
\title{Remove unwanted variation using mixed model}
\usage{
ruv_mixed_model(M, design = NULL, ndups = 1, spacing = 1, block = NULL,
  correlation = NULL, weights = NULL)
}
\arguments{
\item{M}{Data matrix of genes by samples.}

\item{design}{Design matrix of the model fixed effects.}

\item{block}{A vector that uniquely identifies levels of each random
effect.}

\item{correlation}{Consensus correlation coefficient describing the
         average correlation between biological replicates. This is
         usually computed via the duplicateCorrelation() function in
         the limma package.}
}
\description{
This function is a modified version of limma gls.series
function. We changed to gls.series function to include
model residuals that are computed after removing variation
due to the random effect component. Otherwise, no changes was
made to the gls.series algorithm.
}
\examples{
ruv_mixed_model()
}
\keyword{batch}

