% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-DE.r
\name{testDE}
\alias{testDE}
\title{Inter-species divergence likelihood ratio test}
\usage{
testDE(eSet, seqData, species)
}
\arguments{
\item{eSet}{Expression set}

\item{seqData}{A factor vector of sequencing data type.}

\item{species}{A factor vector of species.}
}
\description{
This function tests for divergence between two species
in one molecular phenotype at a time.
}
\examples{
load(file.path(rdadir,"eSetRRP.rda"))
massRes = testDE(eSetRRPall,seqData="protein",species=c("human","chimp"))$massRes
}
\keyword{Humanzee}

