% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/voomWeightsCustomized.r
\name{voomWeightsCustomized}
\alias{voomWeightsCustomized}
\title{Compute voom weights using customized log2-cpm values.}
\usage{
voomWeightsCustomized(log2cpm, lib.size, design)
}
\arguments{
\item{log2cpm}{log2cpm values.}

\item{lib.size}{Library size.}

\item{design}{Experimental design of the data. Required to be an R
design.matrix object}
}
\description{
Compute voom weights using customized log2-cpm values.
}
\examples{
voomWeightsCustomized()
}

