% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/GOtest.R
\name{GOtest}
\alias{GOtest}
\title{Convert ENSG IDs to ENTREZ IDs}
\usage{
GOtest(my_ensembl_gene_universe, my_ensembl_gene_test, pval_cutoff = 0.01,
  ontology = c("BP", "CC", "MF"), conditional.method = F)
}
\arguments{
\item{my_ensembl_gene_universe:}{gene universe}

\item{my_ensembl_gene_test:}{genes of interest}

\item{pval_cutoff:}{cutoff when determining significant enrichment GO terms}

\item{ontology:}{a list of GO ontologies that will be tested in the enrichment anlaysis.
Default: ontology=c("BP","CC","MF").}

\item{pval_cutoff:}{cutoff when determining significant enrichment GO terms.
Default: pval_cutoff = .01.}
}
\value{
hg_over: lists of enrichment test results, each one is a GOHyperGResult
       instance of the GOstats package.
}
\description{
Convert ENSG IDs to ENTREZ IDs
}
\examples{
GOtest()
}

