% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute-cells.r
\name{permute_cells}
\alias{permute_cells}
\title{Permute cell labels}
\usage{
permute_cells(log2counts, grouping_vector, num_permute)
}
\arguments{
\item{log2counts}{log2counts matrix of gene by cells}

\item{grouping_vector}{the grouping vector corresponds to variable of interest}

\item{num_permute}{number of permuted samples}
}
\description{
Permute cell labels
}
\examples{
permuate_cells()
}

