% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute-interaction.r
\name{permute_interact}
\alias{permute_interact}
\title{Wapper function for running interaction model on null data sts}
\usage{
permute_interact(eset_full, datatypes, permute_labels, parallel = FALSE,
  ncores = 4)
}
\arguments{
\item{permute_labels}{}
}
\description{
Wapper function for running interaction model on null data sts
}
\examples{
permute_interact()
}

