% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute-interact-per-gene.R
\name{permute_interact_per_gene}
\alias{permute_interact_per_gene}
\title{Permutation-based test for per gene interaction models}
\usage{
permute_interact_per_gene(eset_full, datatypes, permute_phenotype, n_permute,
  ncores = 4)
}
\arguments{
\item{eset_full}{}

\item{datatypes}{data types included in the analysis}

\item{n_permute}{number of permutation}

\item{ncores}{}

\item{permute_phenotypes}{phenotypes that we would like to shuffle labels between/within}
}
\description{
Written for Sidneys' paper on inter-species comparsions of 
translation regulation. This function performs permutation of sample
lables across species and/or across technology type (RNA, Ribo, Protein)
to compute empirical p-values of interaction models.
}
\examples{
permute_interact_per_gene()
}
\keyword{Humanzee}

