% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testDE.r
\name{testDE}
\alias{testDE}
\title{Inter-species divergence likelihood ratio test}
\usage{
testDE(eSet, seqData, species)
}
\arguments{
\item{eSet}{}

\item{seqData}{}

\item{species}{}
}
\description{
This function tests for divergence between two species
in one molecular phenotype at a time.
}
\examples{
load(file.path(rdadir,"eSetRRP.rda"))
massRes = testDE(eSetRRPall,seqData="protein",species=c("human","chimp"))$massRes
}
\keyword{Humanzee}

